Sprites & Ability Changes
Adds in new sprites made by Ook for Florida Man's abilities.

Decreases the cooldown of Lesser Narcotimancy to make rolling for drugs a bit less painful.

Increases the possible number of chemicals that Lesser Narcotimancy can provide.

The U.S. state of Florida is home to the world’s most dangerous tree – the Manchineel tree. All parts of the tree contain strong toxins.
Mere contact with the sap from this tree can cause blisters on the skin. The tree is also known as “the beach apple” and “little apple of death.