Radio Hotkey
Adds a hotkey (Default Y) which gives players a dropdown menu listing all radio channels available to the user.
Selecting a channel will cause the player to then immediately speak into that channel instead of requiring a key in the text.

Here's a health to the company and one to my lass.
Let's drink and be merry all out of one glass.
Let's drink and be merry, all grief to refrain.
For we may or might never all meet here again.